requirements.txt
crewai
openai
python-dotenv
requests
ollama
pytesseract==0.3.10
python-docx==0.8.11
PyPDF2==3.0.0
pyreadline3  # Replace readline with pyreadline3 for Windows compatibility